requirements.txt
streamlit==1.55.0
altair==6.0.0
psycopg2-binary==2.9.9
pandas==2.2.2